requirements.txt
pillow
imagehash
numpy
fusepy
pathlib